+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/base/adapter/BaseRecyclerViewAdapter.kt
@@ -5,7 +5,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.annotation.LayoutRes
 import androidx.recyclerview.widget.RecyclerView
-import mustafaozhan.github.com.mycurrencies.room.model.Currency
+import mustafaozhan.github.com.mycurrencies.model.Currency
 import mustafaozhan.github.com.mycurrencies.tools.Currencies
 import kotlin.properties.Delegates
