+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/base/api/backend/BackendApiServices.kt
@@ -1,7 +1,7 @@
 package mustafaozhan.github.com.mycurrencies.base.api.backend
 
 import io.reactivex.Observable
-import mustafaozhan.github.com.mycurrencies.main.fragment.model.CurrencyResponse
+import mustafaozhan.github.com.mycurrencies.model.CurrencyResponse
 import mustafaozhan.github.com.mycurrencies.tools.Currencies
 import retrofit2.http.GET
 import retrofit2.http.Query
