+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/settings/SettingsFragmentViewModel.kt
@@ -3,8 +3,8 @@ package mustafaozhan.github.com.mycurrencies.settings
 import mustafaozhan.github.com.mycurrencies.base.BaseViewModel
 import mustafaozhan.github.com.mycurrencies.extensions.insertInitialCurrencies
 import mustafaozhan.github.com.mycurrencies.extensions.removeUnUsedCurrencies
+import mustafaozhan.github.com.mycurrencies.model.Currency
 import mustafaozhan.github.com.mycurrencies.room.dao.CurrencyDao
-import mustafaozhan.github.com.mycurrencies.room.model.Currency
 import mustafaozhan.github.com.mycurrencies.tools.Currencies
 import javax.inject.Inject
